meditation_MVP/requirements_mvp.txt
# Core FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# HTTP Client for MeditationDB API
aiohttp>=3.8.0
httpx>=0.24.0

# Audio processing
librosa>=0.9.0
soundfile>=0.12.1
numpy>=1.21.0

# ML & Vector similarity
scikit-learn>=1.3.0
sentence-transformers>=2.2.2
pandas>=1.5.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.0.0
transformers>=4.21.0

# Optional: Script generation
google-generativeai>=0.3.0
